+++ src/integration/BrowserNativeHostRegistration.h
@@ -1,0 +1,5 @@
+#pragma once
+
+namespace BrowserNativeHostRegistration {
+    void registerHostIfConfigured();
+}